rename on_disk module
