+++ lib/segment/src/index/field_index/geo_index/on_disk_geo_index/lifecycle.rs
